chore: split up resource transfer visitors into independent files
