+++ posthog/models/resource_transfer/visitors/__init__.py
@@ -1,0 +1,23 @@
+from posthog.models.resource_transfer.visitors.action import ActionVisitor
+from posthog.models.resource_transfer.visitors.base import ResourceTransferVisitor
+from posthog.models.resource_transfer.visitors.cohort import CohortVisitor
+from posthog.models.resource_transfer.visitors.dashboard import DashboardVisitor
+from posthog.models.resource_transfer.visitors.dashboard_tile import DashboardTileVisitor
+from posthog.models.resource_transfer.visitors.insight import InsightVisitor
+from posthog.models.resource_transfer.visitors.project import ProjectVisitor
+from posthog.models.resource_transfer.visitors.team import TeamVisitor
+from posthog.models.resource_transfer.visitors.text import TextVisitor
+from posthog.models.resource_transfer.visitors.user import UserVisitor
+
+__all__ = [
+    "ResourceTransferVisitor",
+    "ActionVisitor",
+    "CohortVisitor",
+    "DashboardVisitor",
+    "DashboardTileVisitor",
+    "InsightVisitor",
+    "ProjectVisitor",
+    "TeamVisitor",
+    "TextVisitor",
+    "UserVisitor",
+]
